add tests for logger creation
